Refactor: extract half of setup function
